experimental: rename c-cpp to build_generator
ref: https://github.com/google/oss-fuzz-gen/issues/911
