+++ experimental/build_generator/runner.py
@@ -23,8 +23,8 @@ import sys
 import threading
 from typing import List
 
-import constants
-import templates
+from experimental.build_generator import constants
+from experimental.build_generator import templates
 
 silent_global = False
